+++ articles/flow/tutorials/in-depth-course/_ebook-banner.adoc
@@ -1,0 +1,16 @@
+[.toc-banner-source.skip-search-index]
+--
+image:images/ebook.png[migration assistance]
+
+[.toc-banner-source-text]
+*Download free e-book.* +
+The complete guide is also available in an easy-to-follow PDF format.
+
+[.toc-banner-source-link]
+\https://pages.vaadin.com/en/build-a-modern-web-app-with-spring-boot-vaadin-pdf
+
+[source,html,role=render-only]
+----
+include::{root}/frontend/demo/render-banner.ts[render,persistent]
+----
+--
