+++ articles/flow/tutorials/in-depth-course/components-and-layouts.adoc
@@ -138,10 +138,4 @@ You should see an empty window in the browser, and no errors in the console.
 
 image::images/basics/empty-window.png[empty browser window]
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
