+++ articles/flow/tutorials/in-depth-course/configuring-vaadin-grid.adoc
@@ -168,10 +168,4 @@ The easiest way to do it is to call `grid.getColumns()` and then use `forEach` t
 +
 image::images/grid/contacts-grid-auto-width.png[contact grid with correctly sized columns]
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
