+++ articles/flow/tutorials/in-depth-course/creating-a-component.adoc
@@ -184,10 +184,4 @@ image::images/creating-a-component/main-view-with-form.png[main view with form c
 The visual part of the form is now complete. 
 In the next tutorial, you'll make it functional.
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
