+++ articles/flow/tutorials/in-depth-course/data-binding-and-events.adoc
@@ -211,10 +211,4 @@ image::images/communication/new-contact.png[form with new contact]
 
 In the next chapter, you'll add a second screen to the application and learn how to navigate between views.
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
