+++ articles/flow/tutorials/in-depth-course/database-entities-and-services.adoc
@@ -461,10 +461,4 @@ You need to stop and restart the application to make sure all the new POM depend
 
 In the next chapter, you'll use the backend to populate data into a data grid in the browser.
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
