+++ articles/flow/tutorials/in-depth-course/e2e-testing-with-testbench.adoc
@@ -322,10 +322,4 @@ public void loginAsInvalidUserFails() {
 
 The next chapter covers how to make a production build of the application and deploy it to a cloud platform.
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
