+++ articles/flow/tutorials/in-depth-course/filtering-vaadin-grid.adoc
@@ -129,10 +129,4 @@ You should be able to filter the contacts by entering a term in the text field.
 +
 image::images/filtering/contacts-grid-filtering.png[search feature in contact grid]
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
