+++ articles/flow/tutorials/in-depth-course/forms-and-validation.adoc
@@ -263,10 +263,4 @@ private void validateAndSave() {
 
 In the next tutorial, you'll connect the form to the main view so that the selected contact in the form can be edited.
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
