+++ articles/flow/tutorials/in-depth-course/index.adoc
@@ -97,10 +97,4 @@ You can find the full source code for this guide on https://github.com/vaadin-le
 
 You can find video tutorials of each chapter on the https://www.youtube.com/watch?v=quiQg9bgFs8&list=PLcRrh9hGNallPtT2VbUAsrWqvkQ-XE22h[Vaadin YouTube Channel].
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
