+++ articles/flow/tutorials/in-depth-course/installing-and-offline-pwa.adoc
@@ -161,10 +161,4 @@ image::images/pwa/offline-app.png[custom offline page]
 
 In the next chapter, you will add tests to the application: both unit tests and in-browser tests.
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
