+++ articles/flow/tutorials/in-depth-course/login-and-authentication.adoc
@@ -392,10 +392,4 @@ image::images/login/logout-button.png[Log out button on page]
 You have now built a full-stack CRM application with navigation and authentication. 
 In the next tutorial, you'll learn how to make the application installable on mobile and desktop.
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
