+++ articles/flow/tutorials/in-depth-course/navigation-and-layouts.adoc
@@ -281,10 +281,4 @@ image::images/navigation/dashboard-completed.png[Complete dashboard view]
 
 In the next chapter, you'll secure the application by adding a login screen.
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
