+++ articles/flow/tutorials/in-depth-course/production-deployment.adoc
@@ -176,10 +176,4 @@ You now have all the skills you need to get started building real-life applicati
 Please let us know what you thought of the tutorial series and if you have ideas for other topics you want us to cover. 
 You can reach us on https://twitter.com/vaadin[Twitter].
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
