+++ articles/flow/tutorials/in-depth-course/project-setup.adoc
@@ -182,10 +182,4 @@ Vaadin shares many class names (like Button) with Swing, AWT, and JavaFX.
 
 Now that you have a working development environment, you are ready to start building a web application.
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
