+++ articles/flow/tutorials/in-depth-course/unit-and-integration-testing.adoc
@@ -272,10 +272,4 @@ private Contact getFirstItem(Grid<Contact> grid) {
 You now know how to test the application logic both in isolation with unit tests and by injecting dependencies to test the integration between several components. 
 In the next chapter, the tutorial covers how to test the entire application in the browser. 
 
-[.toc-banner-source]
---
-[source,html,role=render-only]
-----
-include::{root}/frontend/demo/banners/ebook-banner.ts[render]
-----
---
+include::{root}/articles/flow/tutorials/in-depth-course/_ebook-banner.adoc[opts=optional]
