+++ dspublisher/theme/styles/banners.css
@@ -1,8 +1,23 @@
 .toc-banner-source {
-  position: fixed;
-  top: 0;
-  left: 0;
-  visibility: hidden;
-  height: 3px;
-  z-index: -9999;
+  display: none;
+}
+
+.toc-banner {
+  margin-top: 2rem;
+  width: 11.25rem;
+}
+
+.toc-banner img {
+  width: 100%;
+  height: auto;
+  border-radius: 0.25rem;
+}
+
+.toc-banner--content {
+  margin-top: 1rem;
+  line-height: 1.5;
+}
+
+.toc-banner > a::after {
+  content: none !important;
 }
